ci: fix release version

+++ scripts/release.sh
@@ -1,17 +1,16 @@
 set -e
 
-#if [ -z "$1" ]
-#  then
-#    echo "No argument supplied, example: ./scripts/release.sh patch"
-#    exit 1
-#fi
+if [ -z "$1" ]
+  then
+    echo "No argument supplied, example: ./scripts/release.sh patch"
+    exit 1
+fi
 
-
-version="1.0.0"
+version=$(npm version $@ --no-git-tag-version)
 sed -i '' "s/\"version\": \".*\"/\"version\": \"$version\"/" wails.json
 # Create a Go file with the version
 echo "package main\n\n// AppVersion is the current version of the app\nconst AppVersion = \"$version\"" > appversion.go
-#git commit -am "Bump version to $version"
+git commit -am "Bump version to $version"
 rm -rf build/bin
 wails build --platform windows/amd64,darwin/universal,linux/amd64
 npx --yes create-dmg@6.1.0 build/bin/yazu.app build/bin --overwrite || true
